+++ src/rules/no-cycle/__test__/project/overlapping/a.ts
@@ -1,0 +1,5 @@
+import { b } from './b';
+
+export const a = 10;
+
+console.log(b);