+++ src/rules/no-cycle/__test__/project/overlapping/b.ts
@@ -1,0 +1,5 @@
+import { c } from './c';
+
+export const b = 10;
+
+console.log(c);