+++ src/rules/no-cycle/__test__/project/overlapping/c.ts
@@ -1,0 +1,6 @@
+import { a } from './a';
+import { d } from './d';
+
+export const c = 10;
+
+console.log(a, d);