+++ src/rules/no-cycle/__test__/project/overlapping/d.ts
@@ -1,0 +1,5 @@
+import { e } from './e';
+
+export const d = 10;
+
+console.log(e);