+++ src/rules/no-cycle/__test__/project/overlapping/e.ts
@@ -1,0 +1,5 @@
+import { a } from './a';
+
+export const e = 10;
+
+console.log(a);